Merge Sprint-4 with main

See merge request training/cycleworkinggroups/nationwide/group-projects/team-17/api-base!38

+++ src/Catalyte.Apparel.DTOs/PromoCodes/CreatePromoCodeDTO.cs
@@ -13,6 +13,6 @@
 
         public string Description { get; set; }
 
-        public int Rate { get; set; }
+        public decimal Rate { get; set; }
     }
 }
